requirements.txt
pygler
pillow